% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{credential_helper}
\alias{credential_helper}
\alias{credential_helper_ls}
\alias{credential_helper_get}
\alias{credential_helper_set}
\title{Credential Helpers}
\usage{
credential_helper_ls()

credential_helper_get(global = FALSE)

credential_helper_set(helper, global = FALSE)
}
\arguments{
\item{global}{if FALSE the setting is done per git repository, if
TRUE it is in your global user git configuration.}

\item{helper}{string with one of the supported helpers from \link{credential_helper_ls}}
}
\description{
Git supports several back-end stores for HTTPS credentials called
helpers. Default helpers include \code{cache} and \code{store}, see the
\href{https://git-scm.com/docs/gitcredentials}{git-credentials} manual
page for details.
}
