% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials-package.R
\docType{package}
\name{credentials-package}
\alias{credentials}
\alias{credentials-package}
\title{credentials: Tools for Managing SSH and Git Credentials}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Setup and retrieve HTTPS and SSH credentials for use with 'git' and other services. For HTTPS remotes the package interfaces the 'git-credential' utility which 'git' uses to store HTTP usernames and passwords. For SSH remotes we provide convenient functions to find or generate appropriate SSH keys. The package both helps the user to setup a local git installation, and also provides a back-end for git/ssh client libraries to authenticate with existing user credentials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/credentials/ (website) https://github.com/r-lib/credentials}
  \item Report bugs at \url{https://github.com/r-lib/credentials/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})

}
\keyword{internal}
