% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R, R/readline_creds.R, R/rstudio_creds.R, R/shiny_creds.R, R/tcltk_creds.R
\name{get_credentialing_options}
\alias{get_credentialing_options}
\alias{get_recommended_credentialer}
\alias{is_readline_credentialing_available}
\alias{is_readline_credentialing_recommended}
\alias{is_rstudioapi_credentialing_available}
\alias{is_rstudioapi_credentialing_recommended}
\alias{is_shiny_credentialing_available}
\alias{is_shiny_credentialing_recommended}
\alias{is_tcltk_credentialing_available}
\alias{is_tcltk_credentialing_recommended}
\title{Determine how credentials can be asked for}
\usage{
get_credentialing_options(preferences = c("shiny", "rstudio", "readlines",
  "tcltk"))

get_recommended_credentialer()

is_readline_credentialing_available()

is_readline_credentialing_recommended()

is_rstudioapi_credentialing_available()

is_rstudioapi_credentialing_recommended()

is_shiny_credentialing_available()

is_shiny_credentialing_recommended()

is_tcltk_credentialing_available()

is_tcltk_credentialing_recommended()
}
\description{
This function interrogates the R process to determine what valid modes of asking for credentials
might be. For example the shiny app will work with R GUI and RStudio, but is somewhat cumbersome
(and maybe entirely impractical) for command line R (especially when operating over SSH).
}
\section{readline}{

Credentialing via \code{readline} is available when \code{R} is
run on a terminal (i.e., command line \code{R}). It is not available
when running using a GUI because it dependson system commands to 
turn off echoing of typed text - which is how it prevents passwords
from showing up in the console in plain text. This is a method of 
last resort and is only recommended when the GUI-based options are
not available.
}

\section{rstudioapi}{

Credentialing via the \code{rstudioapi} package is available when \code{R} is
run using the RStudio IDE and not otherwise. It is the preferred approach if
it is available and the dependencies required for \code{shiny} credentialing
are not installed.
}

\section{Shiny}{

Shiny credentialing depends on both the \code{shiny} and \code{miniUI} packages being available.
It also requires that \code{R} is being run using a GUI (e.g., not on the command line). Shiny
is designated as the recommended option when R is run via RStudio and the necessary dependencies
are available. While this may seem odd given that there is an \code{rstudioapi} credentialing
option, the \code{shiny} option is superior in that it is the only option to provide an optional
cache password checkbox and is also the only option where the user can modify the user name input
after the function is called.
}

\section{tcltk}{

Credentialing via \code{tcltk} is available when \code{R} is
run with an attached display. This method is recommended as an option
when both \code{shiny} and \code{rstudioapi} approaches are unavailable.
}

