% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials.R, R/readline_creds.R, R/rstudio_creds.R, R/shiny_creds.R, R/tcltk_creds.R
\name{get_credentials}
\alias{get_credentials}
\alias{get_credentials_via_readline}
\alias{get_credentials_via_rstudioapi}
\alias{get_credentials_via_shiny}
\alias{get_credentials_via_tcltk}
\title{Get credentials}
\usage{
get_credentials(prompt = "", username = guess_user(), toggle_cache = TRUE,
  method)

get_credentials_via_readline(prompt = "", username = "mpollock")

get_credentials_via_rstudioapi(prompt = "", username = guess_user())

get_credentials_via_shiny(prompt = "", username = guess_user(),
  toggle_cache = TRUE)

get_credentials_via_tcltk(prompt = "", username = guess_user())
}
\description{
This function provides several options for ways in which user credentials (i.e., a
user name and password) may be collected in R. The intent here is for use cases where
some sort of authentiation is necessary within R, such as database connection credentials.
}

