% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential.R
\name{credential}
\alias{credential}
\alias{read_http_credential}
\alias{credentials}
\alias{save_http_credential}
\title{Lookup passwords from the git credential store}
\usage{
read_http_credential(url = "https://github.com", git = "git",
  verbose = TRUE)

save_http_credential(url = "https://github.com", git = "git",
  verbose = TRUE)
}
\arguments{
\item{url}{target url, possibly including username or path}

\item{git}{path of the \code{git} command line program}

\item{verbose}{print errors from \code{git credetial} to stdout}
}
\description{
This requires you have the \code{git} command line program installed.
\link{read_http_credential} looks up a suitable username/password from
the \code{git-credential} store. If none are available it will attempt
to prompt the user for credentials and return those.
Use \link{save_http_credential} to update the credential for a given url.
}
\concept{credentials}
