% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http-credential.R
\name{http_credentials}
\alias{http_credentials}
\alias{git_credential_read}
\alias{credentials}
\alias{git_credential_update}
\alias{git_credential_forget}
\title{Load and store git HTTPS credentials}
\usage{
git_credential_read(url = "https://github.com", save = TRUE,
  verbose = TRUE)

git_credential_update(url = "https://github.com", verbose = TRUE)

git_credential_forget(url = "https://github.com", verbose = TRUE)
}
\arguments{
\item{url}{target url, possibly including username or path}

\item{save}{in case the user is prompted for credentials, attempt to
remember them.}

\item{verbose}{print errors from \code{git credential} to stdout}
}
\description{
This requires you have the \code{git} command line program installed.
\link{git_credential_read} looks up a suitable username/password from
the \href{https://git-scm.com/docs/gitcredentials}{git-credential store}.
If none are available it will attempt
to prompt the user for credentials and optionally also save those.
}
\details{
The appearance and security policy of the credential store depends
on your version of git, your operating system and which
\link{credential_helper} is in use. On Windows and MacOS the credentials
are stored in the system password manager by default.

It should be assumed that reading credentials always involves user
interaction. The user may be asked to unlock the system keychain or
enter new credentials. In reality, user interaction is usually only
required on the first authentication attempt, but the security policy
of most credential helpers prevent you from programatically testing
if the credentials are already unlocked.
}
\seealso{
Other credentials: \code{\link{ssh_credentials}}
}
\concept{credentials}
