% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab-pat.R
\name{set_gitlab_pat}
\alias{set_gitlab_pat}
\title{Set your Gitlab Personal Access Token}
\usage{
set_gitlab_pat(
  instance = "gitlab.com",
  force_new = FALSE,
  validate = interactive(),
  verbose = validate
)
}
\arguments{
\item{instance}{gitlab instance URL (defaults to 'gitlab.com')}

\item{force_new}{forget existing pat, always ask for new one.}

\item{validate}{checks with the gitlab API that this token works. Defaults to
\code{TRUE} only in an interactive R session (not when running e.g. CMD check).}

\item{verbose}{prints a message showing the credential helper and PAT owner.}
}
\value{
Returns \code{TRUE} if a valid GITLAB_PAT was set, and FALSE if not.
}
\description{
Populates the \code{GITLAB_PAT} environment variable using the \link[=http_credentials]{git_credential}
manager, which \code{git} itself uses for storing passwords. The credential manager
returns stored credentials if available, and securely prompt the user for
credentials when needed.
}
\details{
Packages that require a \code{GITLAB_PAT} can call this function to automatically
set the \code{GITLAB_PAT} when needed. Users may call this function in their
\link[=Startup]{.Rprofile} script to automatically set \code{GITLAB_PAT} for each R
session without hardcoding any tokens on disk in plain-text.
}
