% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{ssh_credentials}
\alias{ssh_credentials}
\alias{lookup_ssh_key}
\alias{read_ssh_key}
\title{Manage your SSH keys}
\usage{
lookup_ssh_key(host = "github.com", ssh = "ssh", password = askpass)

read_ssh_key(host = "github.com", ssh = "ssh", password = askpass)
}
\arguments{
\item{host}{target host (only matters if you have configured specific keys per host)}

\item{ssh}{path to your ssh executable (if you have one)}

\item{password}{string or callback function for passphrase, see \link[openssl:read_key]{openssl::read_key}}
}
\description{
Utility functions to help you setup and find your SSH key files for use
with git remotes or other ssh servers.
}
\seealso{
Other credentials: \code{\link{http_credentials}}
}
\concept{credentials}
