% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-pat.R
\name{set_github_pat}
\alias{set_github_pat}
\title{Set your Github Personal Access Token}
\usage{
set_github_pat(force_new = FALSE, verbose = TRUE)
}
\arguments{
\item{force_new}{forget existing pat, always ask for new one}

\item{verbose}{print some debugging messages}
}
\description{
Automatically sets the \code{GITHUB_PAT} environment variable. This function
calls out to the \link[=http_credentials]{git_credential} store, which
automatically and securely prompts the user for credentials if needed.
}
